requirements.txt
analytics-python==1.2.9
appnope==0.1.0
arrow==0.12.1
astroid==1.6.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
binaryornot==0.4.4
boto==2.41.0
boto3==1.4.7
botocore==1.7.48
certifi==2018.4.16
chardet==3.0.4
click==4.0
configparser==3.5.0
contextlib2==0.5.5
cookiecutter==1.5.1
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.6
django-braces==1.13.0
django-chartit==0.2.9
django-cors-headers==2.2.0
django-facebook==6.0.2
django-fine-uploader-s3==0.1.14
django-log-request-id==1.3.2
django-oauth==1.1
django-oauth-toolkit==0.10.0
django-smtp-ssl==1.0
django-storages==1.4.1
django-suit==0.2.20
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.2.0
google==2.0.1
googleapis-common-protos==1.5.3
grpc-django==0.3.dev0
grpc-pytools==0.1.2
grpcio==1.12.1
grpcio-tools==1.12.1
idna==2.7
ipython==5.7.0
ipython-genutils==0.2.0
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
jsonpatch==1.23
jsonpointer==2.0
jsonschema==2.6.0
lazy-object-proxy==1.3.1
lazyme==0.0.20
markdown2==2.3.5
MarkupSafe==1.0
marshmallow==2.15.3
mock==2.0.0
oauth==1.0.1
oauthlib==1.0.3
otpauth==1.0.1
pathlib2==2.3.2
pbr==4.0.4
pexpect==4.6.0
pickleshare==0.7.4
poyo==0.4.1
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
Pygments==2.2.0
pykerberos==1.2.1
python-dateutil==2.7.3
Python-EasyConfig==0.1.6
Python-RESTArt==0.1.5
pytz==2018.4
PyYAML==3.11
raven==6.2.1
requests==2.19.1
requests-kerberos==0.11.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
scandir==1.7
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snapshottest==0.5.1
suit==2.0.2
swagger-parser==0.1.7
swagger-spec-validator==2.0.2
termcolor==1.1.0
traitlets==4.3.2
twython==3.4.0
Unidecode==1.0.22
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
whichcraft==0.4.1
wrapt==1.10.11
